done implementing welcome page
